refactor: split reporter tests into relevant files and reduce test runtime
